Refactor code in response to feedback
