+++ storeapi/models/__init__.py
@@ -1,0 +1,9 @@
+# from flask import Blueprint
+
+# api = Blueprint(
+#     'site',
+#     __name__,
+#     template_folder='templates',
+#     static_folder='static'
+# )
+
