+++ storeapi/views/__init__.py
@@ -1,0 +1,28 @@
+# from flask import Flask
+# from flask_jwt_extended import (JWTManager, create_access_token,
+#                                 get_jwt_identity, jwt_required)
+
+# # from app import create_app
+# from storeapi.models.product_model import Product
+# from storeapi.models.sales_model import Sale
+# from storeapi.views.auth_views import app
+# from storeapi.views.product_views import app
+# from storeapi.views.sale_views import app
+
+# from app import create_app
+# app = create_app('development')
+
+# app = Flask(__name__)
+
+
+
+# from flask import Blueprint
+
+# api = Blueprint(
+#     'site',
+#     __name__,
+#     template_folder='templates',
+#     static_folder='static'
+# )
+
+
