+++ tests/test_sales.py
@@ -1,0 +1,48 @@
+from flask import Flask,json
+
+from tests import BaseTestCase
+from flask_jwt_extended import JWTManager, create_access_token
+
+
+class SalesTestCase(BaseTestCase):
+
+    def test_add_sale_order(self):
+        with self.app.app_context():
+            token = create_access_token('attendant')
+            headers = {'Authorization': f'Bearer {token}'}
+            response = self.test_client.post(
+                '/api/v1/sales',
+                headers=headers,
+                content_type='application/json'
+            )
+            return(response.status)
+        self.assertEqual(response.status_code,200)
+        self.assertIn(
+            'You have successfully created a sale order', str(response.data)
+        )
+    def test_fetch_single_sale_order(self):
+        with self.app.app_context():
+            token = create_access_token('admin')
+            headers = {'Authorization': f'Bearer {token}'}
+            response = self.test_client.get(
+                '/api/v1/sales/1',
+                headers=headers,
+                content_type='application/json'
+            )
+            return(response.status)
+        self.assertEqual(200, response.status_code)
+       
+        self.assertEqual(response.status_code,200)
+
+    def test_fetch_all_sale_orders(self):
+        
+        with self.app.app_context():
+            token = create_access_token('admin')
+            headers = {'Authorization': f'Bearer {token}'}
+            response = self.test_client.get(
+                '/api/v1/sales',
+                headers=headers,
+                content_type='application/json'
+            )
+            return(response.status)
+        self.assertEqual(200, response.status_code)
